Adds measure of uncertainty for AML-based  fundamental matrix estimate
Adds an analytic expression for the covariance matrix of the AML-based fundamental matrix estimate. The uncertainty of the fundamental matrix estimate also accounts for the post-hoc  SVD-based rank-2 enforcement that must be applied to ensure a bonafide fundamental matrix. 

This commit includes some dead code which I have temporarily commented out as I am in the process of refactoring and evaluating the performance of the code using StaticArrays.